Initial commit of Radio Automation package.
There is a lot of work here to setup the core functionality. Project uses Catel for MVVM assistance, NAudio for playing audio. The goal here is to replace another Radio Automation package I was using that was frozen in time and had some bugs that were time consuming to work around. I also don;t have the need to for a ful lRadio Automation functionality in the true sense of a radio station so this will just contain the features I need.